refactor operator console page families
